# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.createAPIKeyRequest:
    properties:
      expires_in_days:
        type: integer
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
...
  handlers.createRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
    type: object
...
  handlers.inviteRequest:
    properties:
      email:
        type: string
      role_id:
        type: string
    type: object
...
  handlers.updateMemberRequest:
    properties:
      role_id:
        type: string
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@strata.dev
    name: Strata Support
  description: ERP-CRM Hybrid API — direct competitor to Odoo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Strata API
  version: 0.1.0
paths:
...
  /api/v1/org/api-keys:
    post:
      consumes:
      - application/json
      description: Generates a new API key for machine-to-machine integrations. Requires
        `apikeys.manage` permission.
      operationId: createAPIKey
      parameters:
      - description: API key payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - Organizations
  /api/v1/org/invitations:
    post:
      consumes:
      - application/json
      description: Sends an invitation to join the organization. Requires `users.invite`
        permission.
      operationId: inviteMember
      parameters:
      - description: Invitation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.inviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Invite team member
      tags:
      - Organizations
  /api/v1/org/members/{member_id}:
    delete:
      description: Soft-deletes a member by setting is_active = false. Requires `users.manage`
        permission.
      operationId: deactivateMember
      parameters:
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Deactivate organization member
      tags:
      - Organizations
    patch:
      consumes:
      - application/json
      description: Updates a member's role or other admin-managed fields. Requires
        `users.manage` permission.
      operationId: updateMember
      parameters:
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Update organization member
      tags:
      - Organizations
  /api/v1/org/members/{member_id}/remove:
    post:
      description: |-
        Removes a member from the organization's member list without deleting the user account. Requires `users.manage` permission.
        Permanently removes a member from the organization by deleting the membership row. The user account is preserved. Requires `users.manage` permission.
      operationId: removeMember
      parameters:
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Remove organization member
      tags:
      - Organizations
  /api/v1/org/roles:
    post:
      consumes:
      - application/json
      description: Creates a new role with assigned permissions. Requires `rbac.manage`
        permission.
      operationId: createRole
      parameters:
      - description: Role payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Create dynamic role
      tags:
      - Organizations
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access (e.g. fleet telematics ingestion)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...